Remove service impl to just be service
