+++ universal-application-tool-0.0.1/app/services/role/RoleService.java
@@ -1,18 +1,37 @@
 package services.role;
 
+import static com.google.common.collect.ImmutableSet.toImmutableSet;
+
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Sets;
+import javax.inject.Inject;
 import models.Account;
+import repository.UserRepository;
+import services.program.ProgramDefinition;
 import services.program.ProgramNotFoundException;
+import services.program.ProgramService;
 
 /** A service for reading and updating data related to system roles. */
-public interface RoleService {
+public class RoleService {
+
+  private final ProgramService programService;
+  private final UserRepository userRepository;
+
+  @Inject
+  public RoleService(ProgramService programRepository, UserRepository userRepository) {
+    this.programService = programRepository;
+    this.userRepository = userRepository;
+  }
 
   /**
    * Get a set of {@link Account}s that have the role {@link auth.Roles#ROLE_UAT_ADMIN}.
    *
    * @return an {@link ImmutableSet} of {@link Account}s that are UAT admins.
    */
-  ImmutableSet<Account> getUatAdmins();
+  public ImmutableSet<Account> getUatAdmins() {
+    // TODO(cdanzi): implement this method
+    return ImmutableSet.of();
+  }
 
   /**
    * Promotes the set of accounts (identified by email) to the role of {@link
@@ -23,6 +42,13 @@ public interface RoleService {
    * @param programId the ID of the {@link models.Program} these accounts administer
    * @param accountEmails a {@link ImmutableSet} of account emails to make program admins
    */
-  void makeProgramAdmins(long programId, ImmutableSet<String> accountEmails)
-      throws ProgramNotFoundException;
+  public void makeProgramAdmins(long programId, ImmutableSet<String> accountEmails)
+      throws ProgramNotFoundException {
+    ProgramDefinition program = programService.getProgramDefinition(programId);
+    // Filter out UAT admins from the list of emails - a UAT admin cannot be a program admin.
+    Sets.difference(
+            accountEmails,
+            getUatAdmins().stream().map(Account::getEmailAddress).collect(toImmutableSet()))
+        .forEach(email -> userRepository.addAdministeredProgram(email, program));
+  }
 }
